Suggested refactor: extract shared epochs/corepoints handling into PairwiseEpochCorepointsBase
